+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Rename `extra` management script runner to `script`
 - Remove support of environment configuration files (`config.{development,production}.json`), set `CONFIG_FILE` instead!
 - Don't write config file if not existing or missing key `namespace`, please generate by hand!
+- Use status code 201 only when data has been created
 
 ## 2.4.2